+++ include/nau7802.h
@@ -86,10 +86,19 @@ int nau7802_read(i2c_master_dev_handle_t i2c, int32_t* val);
 // on success, val will hold some value less than scale.
 int nau7802_read_scaled(i2c_master_dev_handle_t i2c, float* val, uint32_t scale);
 
+// disable or enable the bandgap chopper. it is enabled by default.
+int nau7802_set_bandgap_chop(i2c_master_dev_handle_t i2c, bool enabled);
+
 // the device can be put into an extreme powered-down mode, which shuts
 // down the entire analog portion of the part. it must be brought out of
 // this mode before reads can be performed again. pass true to enter
 // the powered down state, or false to leave it.
 int nau7802_set_deepsleep(i2c_master_dev_handle_t i2c, bool powerdown);
+
+// the DRDY pin can either indicate that there is a conversion ready to be
+// read (the default), or it can export the clock. pass true to export the
+// clock being used (depends on OSCS), false to reestablish the default
+// behavior of indicating data readiness.
+int nau7802_export_clock(i2c_master_dev_handle_t i2c, bool clock);
 
 #endif